+++ Unity.Collections/NativeParallelHashSet.cs
@@ -72,6 +72,11 @@ namespace Unity.Collections
         }
 
         /// <summary>
+        /// The maximum number of elements this type of container can hold.
+        /// </summary>
+        public const int MaxCapacity = NativeParallelHashMap<T, bool>.MaxCapacity;
+
+        /// <summary>
         /// Whether this set has been allocated (and not yet deallocated).
         /// </summary>
         /// <value>True if this set has been allocated (and not yet deallocated).</value>